Ark KPart: restore mismatch of componentname and plugin id
By history Ark stores the xmlgui ui.rc file for the part in the same
subdir as the program, ark/. And set the comppnent name as needed to "ark"
via the KAboutData instance for the KPart.
At the same time the plugin id has been "arkpart", as derived from the
library name due to the absence of an explicit X-KDE-PluginInfo-Name entry
in the desktop file.

With the change to the KPluginMetaData API of KParts/KPluginLoader and
the new entries added to the desltop file based on the KAboutData values,
that mismatch was resolved by accident. With "ark" now also the plugin id,
the look up the kpart by the program though now fails, due to the condition
    metaData.pluginId() == QLatin1String("arkpart")